reuse the screens file instead of importing the entire tw config file
