+++ src/lib/types.ts
@@ -23,10 +23,10 @@ import type { SimulatorNav } from "@/components/Simulator/interfaces"
 import allQuizData from "@/data/quizzes"
 import allQuestionData from "@/data/quizzes/questionBank"
 
+import { screens } from "./utils/screen"
 import { WALLETS_FILTERS_DEFAULT } from "./constants"
 
 import { layoutMapping } from "@/pages/[...slug]"
-import twConfig from "@/styles/config"
 
 // Credit: https://stackoverflow.com/a/52331580
 export type Unpacked<T> = T extends (infer U)[] ? U : T
@@ -858,4 +858,4 @@ export type EventCardProps = {
   imageUrl?: string
 }
 
-export type BreakpointKey = keyof typeof twConfig.theme.screens
+export type BreakpointKey = keyof typeof screens
