--- src/styles/config.ts
@@ -1,7 +1,0 @@
-import resolveConfig from "tailwindcss/resolveConfig"
-
-import tailwindConfig from "../../tailwind.config"
-
-const config = resolveConfig(tailwindConfig)
-
-export default config
