--- tailwind/screens.ts
@@ -1,9 +1,0 @@
-import type { ScreensConfig } from "tailwindcss/types/config"
-
-export const screens = {
-  sm: "480px",
-  md: "768px",
-  lg: "992px",
-  xl: "1280px",
-  "2xl": "1536px",
-} satisfies ScreensConfig
